requirements.txt
# --- Async & Networking ---
aiohttp>=3.9.0
websockets>=12.0
requests>=2.31.0

# --- Polymarket ---
py-clob-client

# --- HFT Performance Boosters ---
uvloop>=0.19.0; sys_platform != 'win32'
orjson>=3.9.0

# --- Data, Math & Backtesting ---
numpy>=1.26.0
pydantic>=2.0.0
pandas>=2.0.0

# --- UI & Config ---
rich>=13.7.0
python-dotenv>=1.0.0
